+++ on-my-way/src/station/components.rs
@@ -3,5 +3,5 @@ use bevy::prelude::*;
 pub struct FuelStation {
     pub direction: Vec2,
     pub speed: f32,
-    pub fuel_amount: u32,
+    pub fuel_amount: f32,
 }
